Create File Handling modified code
To modify the code to append new data to the file rather than overwriting previous input, you can change the file mode from `"w"` (write mode) to `"a"` (append mode) when opening the file for writing.

With these changes, new data entered by the user will be appended to the existing content of the input file, and when reading from the input file and writing to the output file, the output file will also be appended with the contents of the input file. This ensures that previous input and output are retained and not overwritten.